Alter position of title for category bar graph
Change width of the title of the bar graph so that it is more aligned with the graph.
QADP-46: As a Data Manager I can see the total number of PDs that have occurred at site across all the studies, so that I can monitor how many PDs are occurring at site.