Add clarity to upload system
Change message colour to green if file is uploaded successfully.
Remove message when new file is chosen.
Remove selected file name when upload is clicked a second time.
QADP-139: when press upload remove the selected file, so that when upload is pressed again it would not upload the file a second time.