Separate components onto different pages and implement tab component
Separate each component on the admin page to be on a separate page. Also
includes changing url admin -> administration-page and using tab
component in shared module.
QADP-87: As a Administrator I can see a team management page, so that I
can create teams and assign users to one or multiple teams so that a
user can use the application.