Write api to retrieve data for table
Add Lombok to dependencies to reduce boilerplate Java code.
Create table based on label that is clicked on 2nd graph.
QADP-63: As a Data Manager I want to select a PD coded term (DVDECOD) from the graph QADP-62 and see a table with the sponsor descriptor (DVSPONDES) and NLP assigned DVDECOD and DVCAT at team level, so that I can see the granularity of the data.