Handle issue with updating labels for 2nd graph
Update 2nd graph correctly to get correct data for table.
QADP-63: As a Data Manager I want to select a PD coded term (DVDECOD) from the graph QADP-62 and see a table with the sponsor descriptor (DVSPONDES) and NLP assigned DVDECOD and DVCAT at team level, so that I can see the granularity of the data.